Update `filecoin-cxx` deps, mainly `multihash`
This change is a first pass in getting the dependecies for
`filecoin-cxx` updated, to eventually be able to drop the unmaintained
crate `proc-macro-error`.

This change updates `multihash` to `v0.16.3`->`v0.18.1`. In a follow up
PR, the goal will be to deploy `v0.19.1`, which drops the unwanted crate
as a transient dependency.